rename: 도메인 내 패키지 세분화
