+++ src/main/java/blackjack/domain/card/CardNumber.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum CardNumber {
     A("A", 1),
