+++ src/main/java/blackjack/domain/gamer/Gamer.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.gamer;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardGroup;
 
 public class Gamer {
     private static final String NAME_LENGTH_ERROR_MESSAGE = "이름은 최소 1글자, 최대 6글자로 입력해야 합니다.";
