+++ src/main/java/blackjack/domain/gamer/Player.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.gamer;
 
 import java.util.ArrayList;
 import java.util.List;
