+++ src/main/java/blackjack/domain/result/DealerResult.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 import java.util.Map;
