+++ src/main/java/blackjack/domain/result/Match.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 import java.util.Arrays;
 import java.util.Optional;
