+++ src/main/java/blackjack/domain/result/PlayerResult.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 import java.util.Map;
