+++ src/test/java/blackjack/domain/PlayerTest.java
@@ -2,6 +2,8 @@ package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.domain.card.CardShape;
+import blackjack.domain.gamer.Player;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
